requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
deprecation==2.1.0
Flask==2.3.2
frozendict==2.4.6
frozenlist==1.5.0
gotrue==2.11.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
multidict==6.1.0
multitasking==0.0.11
numpy==2.2.1
packaging==24.2
pandas==2.2.3
peewee==3.17.8
platformdirs==4.3.6
postgrest==0.18.0
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
realtime==2.0.6
requests==2.31.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.20
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
webencodings==0.5.1
websockets==13.1
Werkzeug==3.1.3
yarl==1.18.3
yfinance==0.2.52
